Pursuing cleaning with store things as a package and removing 'tcut'.
