+++ .pre-commit-config.yaml
@@ -87,7 +87,8 @@ repos:
             --per-file-ignores=
                 tests/*.py:D103
                 oups/__init__.py:F401
-                oups/writer.py:C901
+                oups/store/__init__.py:F401
+                oups/store/writer.py:C901
                 oups/streamagg/__init__.py:F401
                 oups/streamagg/jcumsegagg.py:C901
                 oups/streamagg/segmentby.py:C901
