+++ oups/__init__.py
@@ -5,9 +5,9 @@ Created on Wed Dec  1 18:35:00 2021.
 @author: yoh
 
 """
-from .collection import ParquetSet
-from .indexer import is_toplevel
-from .indexer import sublevel
-from .indexer import toplevel
+from .store import ParquetSet
+from .store import is_toplevel
+from .store import sublevel
+from .store import toplevel
 from .streamagg import by_x_rows
 from .streamagg import streamagg
