+++ oups/store/defines.py
