+++ tests/test_streamagg/__init__.py
@@ -5,5 +5,3 @@ Created on Wed Dec  1 18:35:00 2021.
 @author: yoh
 
 """
-
-TEST_DATA = "test-data"
