+++ tests/test_streamagg/test_streamagg_multi.py
@@ -23,11 +23,11 @@ from pandas.core.resample import TimeGrouper
 from oups import ParquetSet
 from oups import streamagg
 from oups import toplevel
+from oups.store.writer import MAX_ROW_GROUP_SIZE
 from oups.streamagg.jcumsegagg import FIRST
 from oups.streamagg.jcumsegagg import LAST
 from oups.streamagg.segmentby import by_x_rows
 from oups.streamagg.streamagg import _setup
-from oups.writer import MAX_ROW_GROUP_SIZE
 
 
 # from pandas.testing import assert_frame_equal
